+++ src/backend/job-manage/service-job-manage/src/main/java/com/tencent/bk/job/manage/service/globalsetting/GlobalSettingsService.java
@@ -32,14 +32,11 @@ import com.tencent.bk.job.manage.model.web.request.globalsetting.HistoryExpireRe
 import com.tencent.bk.job.manage.model.web.request.notify.ChannelTemplatePreviewReq;
 import com.tencent.bk.job.manage.model.web.request.notify.ChannelTemplateReq;
 import com.tencent.bk.job.manage.model.web.request.notify.NotifyBlackUsersReq;
-import com.tencent.bk.job.manage.model.web.request.notify.SetAvailableNotifyChannelReq;
 import com.tencent.bk.job.manage.model.web.vo.globalsetting.AccountNameRulesWithDefaultVO;
 import com.tencent.bk.job.manage.model.web.vo.globalsetting.FileUploadSettingVO;
-import com.tencent.bk.job.manage.model.web.vo.globalsetting.NotifyChannelWithIconVO;
 import com.tencent.bk.job.manage.model.web.vo.globalsetting.PlatformInfoVO;
 import com.tencent.bk.job.manage.model.web.vo.globalsetting.PlatformInfoWithDefaultVO;
 import com.tencent.bk.job.manage.model.web.vo.notify.ChannelTemplateDetailWithDefaultVO;
-import com.tencent.bk.job.manage.model.web.vo.notify.ChannelTemplateStatusVO;
 import com.tencent.bk.job.manage.model.web.vo.notify.NotifyBlackUserInfoVO;
 
 import java.util.List;
@@ -49,10 +46,6 @@ public interface GlobalSettingsService {
     Boolean isNotifyChannelConfiged(String tenantId);
 
     Boolean setNotifyChannelConfiged(String tenantId);
-
-    List<NotifyChannelWithIconVO> listNotifyChannel(String username);
-
-    Integer setAvailableNotifyChannel(String username, SetAvailableNotifyChannelReq req);
 
     List<NotifyBlackUserInfoVO> listNotifyBlackUsers(String username, Integer start, Integer pageSize);
 
@@ -92,8 +85,6 @@ public interface GlobalSettingsService {
 
     ChannelTemplateDetailWithDefaultVO getChannelTemplateDetail(String username, String channelCode,
                                                                 String messageTypeCode);
-
-    List<ChannelTemplateStatusVO> listChannelTemplateStatus(String username);
 
     String getDocCenterBaseUrl();
